pyproject.toml
# All markdown in this file was auto-generated by Poetry, the python packaging tool. https://python-poetry.org/docs/pyproject/

[tool.poetry]
name = "340_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.3.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
databases = "^0.8.0"
aiomysql = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
